+++ presentation/src/main/java/team/retum/jobis_android/viewmodel/recruitment/RecruitmentViewModel.kt
@@ -169,12 +169,12 @@ internal class RecruitmentViewModel @Inject constructor(
 }
 
 data class RecruitmentUiModel(
-    val recruitId: Int,
+    val recruitId: Long,
     val companyName: String,
     val companyProfileUrl: String,
-    val trainPay: Int,
+    val trainPay: Long,
     val military: Boolean,
-    val totalHiring: Int,
+    val totalHiring: Long,
     val jobCodeList: String,
     var bookmarked: Boolean,
 )
